Conslidated some of the shared detail
- Consolidate the fixedqueue into the shared detail
- move the sha1 file into the shared deatail folder
