+++ src/detail/shared/sha1.cpp
