+++ src/detail/shared/sha1.h
